youtube_chbot/requirements.txt
fastapi
uvicorn
youtube-transcript-api
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
faiss-cpu
sentence-transformers
python-dotenv

